requirements.txt
flask
flask-cors
google-genai
sympy
python-dotenv
gunicorn
markdown
bleach
